agent: Trace+Config services implemented
* all: rewrite from groundup
Due to the complexity of setting up repositories, it has
been advised that we reuse this repository. However, since
the agent spec has radically changed and is being
implemented from the ground up, we are rewriting this
code hence the deletions.

* Trace & Config services implemented along with
hooks that will be used by the frontend/consumer
of the agent itself. Please see the tests for
how these hooks are used or just read the godoc
for the various `AgentOption`s.

* .travis.yml: upgrade to Go1.11

Fixes #18