plugins hooked into cmd/ocagent
* Also added a canonical example as well as
regression test to ensure that plugins
work as expected.

This change finishes up the work started in #46.

* plugins_test: skip TestCreatePlugin

On Travis CI, trying to retrieve the same version of Go
as is being used at runtime is quite the hassle. Since
this test is a nice-to-have and more of an integration
test that runs properly locally, skip it for now until
we have enough time to expend investigating Travis CI e.g.
    https://travis-ci.org/census-instrumentation/opencensus-service/builds/438157975

Fixes #47