interceptor: make endToEnd test an example test
Previously:
    TestNewInterceptor_endToEnd

served the purpose of demoing how to make an "agent"
with the OpenCensus interceptor. It just spams stdout
with output and doesn't perform any assertions
nor serve any other purpose other than demoing.
As we add other interceptors, we'll put them in
here. Instead make it an example test that compiles
and will serve as the blueprint towards making
cmd/ocagent modifications.