Start adding pipeline to collector
This is the basis to do the work for actually implementing OC collector
in the short run. At first enable receivers:

1. OC Receiver
2. Jaeger
3. Zipkin

The queue processor is a no-op at this stage and should be the next step
added. As previously indicated before there is some duplication of code
with the agent this should be reduced as we implement the interfaces and
project coalesce around certain patterns.