Fix nil Node and allow spans without ip address on LocalEndpoint
The Zipkin exporter was failing if the Node of the batch to be exported
was nil (trying to access an attribute property).

If there was no ip address (either v4 or v6) the call to openzipking.NewEndpoint
was returning error and causing the span to be dropped. This was more restrictive
than the Zipkin backend that allows spans without LocalEndpoint or any of its
fields.

Collateral: updated the tests to include such cases, and changed the config
consistency check to allow usage of localhost to receive and export as long
as they are in different ports.