Adds Batch Queue of Ids (for tail-sampling)
Adding one of the types used to implement tail-based sampling.

This type will be used to create the batches that are periodically
processed while running tail-based sampling. The idea is that
receivers will enqueue ids concurrently in a batch while from time
to time a batch will be dequeued for sampling policy evaluation.

The queue of batches behaves in FIFO fashion and the owner of the type
is in charge of closing each batch by calling dequeue (which returns the
oldest batch in the queue while also starting a new batch).