Convert to Jaeger proto through Thrift
This commit replaces an implementation of OC Proto to Jaeger Proto that
was very identical to OC Proto to Jaeger Thrift with a translator that
converts to OC to Thrift first and then converts Thrift to Jaeger Proto.

This should cut down the cost to maintain two almost identical
implementations for some extra runtime performance hit.