Make TestZipkinExportersFromYAML_roundtripJSON reliable
The test depends on the 3 spans from the test data to arrive as a single batch. However, the zipkin exporter does not exposes this fine control of the zipkin reporter used to send the spans. In principle we could expose an option to adjust zipkin reporter batch size but that still requires the test to coordinate the send/receive of the batch. A simpler solution is to implement a mock of the zipkin reporter that behaves synchronously, avoiding any synchronization issue. This also avoids more API surface on the product code.