Fix flaky test for Zipkin LOCAL_COMPONENT handling
Internally the function being tested uses a map to generate the slice with the translation results, however, the order is not deterministic and some runs fail. There are just two elements on the translated results but to make clear the expectation of the test this change uses sort.Slice function to achieve the desired order.