Implement receivers factory
Implement the receivers factory by adding a template type that can be used by receivers in general. Small changes were made to the factory interface to better cover the usage in the actual code.

Added factory for ZipkinScribe as an example, and showed its usage from collector (did not generalize the others but with this internal/collector/receivers can be removed).