Make all receivers take "next" on creation
This is required in order to move the code to use factories for the receivers. The "next" parameter will be passed to the factory and then to the receiver constructor, this way Start[Trace|Metrics]Reception doesn't need the next parameter anymore. Since typically the Start* methods launch goroutines took the opportunity to add a channel to report asynchronous errors.

Opportunistic change: renamed WrapWithSpanSink and related variables.

Remaing items from previous commit