Implement Configuration V2 loading and validation
This commit is the first in the series of planned commits that implement
Configuration V2 support. For full support of config V2 I plan to submit
several additional PRs after this one is reviewed and merged.

List of changes in this commit:

- Added configuration models
- Implemented generic parts of configuration loading and validation code
- Added global factory registry for receivers, exporters and options
- Support single- and multi- protocol receivers

All new code is only active in tests, it is not reachable from main code.

The next planned PR is the implementation of configuration loading
for specific receivers/exporters/options.

Testing done: automated tests.

To reviewers: this is my first PR against this repo. Any improvement
suggestions are highly welcome.

For reference, the Configuration Format V2 design doc:
https://docs.google.com/document/d/1GWOzV0H0RTN1adiwo7fTmkjfCATDDFGuOB4jp3ldCc8/edit#