Implement Configuration V2 models
This commit adds configuration models and is the first in the series of
planned commits that implement Configuration V2 support. For full support
of config V2 I plan to submit several additional PRs after this one is
reviewed and merged.

The next planned PR is the implementation of V2 configuration loading.

For reference, the Configuration Format V2 design doc:
https://docs.google.com/document/d/1GWOzV0H0RTN1adiwo7fTmkjfCATDDFGuOB4jp3ldCc8/edit#