Fix "vet" target in makefile
Previously "vet" target was coded to fail if govet produced any output,
which is not a correct assumption. Govet can produce output even in
case of success [1]. The correct way to know if govet is successful is
to rely on exit code, which is exactly what is done now after this change.

[1] - see for example https://travis-ci.org/census-instrumentation/opencensus-service/builds/536550187