Add trace head-sampling
This is to complete the sampling feature since OC Service already offers tail-sampling. This is implemented using an extra hashing as an attempt to avoid bias from trace ID generation and also to correctly spread traces for backends that also do hashing.

Tests to enforce hash of different lengths.

Make hashing seed configurable