Add support for specifying service name in config
The datadog exporter will currently drop the service name which is
specified in the spans. See https://github.com/census-instrumentation/opencensus-service/issues/198
for more information.

This aligns the datadog exporter with jaeger, wavefront, and zipkin.
They already have this option.

A better solution would be to add support for multiple services, as many
of the other exporters have, but that is a bit over the level of
contribution I can do at the moment.