receiver/opencensus/ocmetrics: ignore io.EOF on JSON upload
When the grpc-gateway was added to this receiver/opencensus
Trace receiver in PR #270, we recognized that browsers
and other sources that send JSON in one burst won't stream
so would encounter an io.EOF. This PR makes the metrics
receiver replicates that behavior.

We'll get it this PR also ported over to the OpenTelemtry
service.

Fixes #610