receiver: add UNIX Domain Socket
Brings in the source code from opencensus-php/daemon
that implements the UNIX Domain Socket receiver and
performs aggregation on the daemon before forwarding
it to the ocagent via gRPC. The overall design is
present at
https://docs.google.com/document/d/1CRiRq_wpzOuG9VKM_eaImcrS12Iie2V7LePnH9AwclU/

In follow-up commits, we'll then start plumbing it to
work directly with the agent. The first step is bringing
in the source code.